+++ zlgcan/src/api/linux/usbcanfd.rs
@@ -157,7 +157,7 @@ impl ZCanApi for USBCANFDApi<'_> {
     fn init_can_chl(&self, context: &mut ZChannelContext, cfg: &ChannelConfig) -> Result<(), CanError> {
         let (dev_type, dev_idx, channel) = (context.device_type(), context.device_index(), context.channel());
         let cfg_ctx = CanChlCfgContext::new()?;
-        let bc_ctx = cfg_ctx.0.get(&dev_type.to_string())
+        let bc_ctx = cfg_ctx.0.get(&(dev_type as u32).to_string())
             .ok_or(CanError::InitializeError(
                 format!("device: {} is not configured in {}", dev_type, BITRATE_CFG_FILENAME)
             ))?;
